requirements.txt
web3==5.31.1
python-dotenv==0.19.2
requests==2.27.1